Setting combined cross-sections
